+++ CHANGES.md
@@ -3,7 +3,7 @@
 * Released on ?
 * Consistently use camelCase in API
 * No more relies on using eval()
-* Depends on phpmyadmin/simple-math for calculations
+* Depends on symfony/expression-language for calculations
 
 # Version 1.2